+++ content/script/educlient/edu.apiclient.js
@@ -93,6 +93,7 @@ edu.apiclient = {
 				template: 			jQuery('.eduadmin-courselistoptions').data('template'),
 				numberofevents:		jQuery('.eduadmin-courselistoptions').data('numberofevents'),
 				fetchmonths:		jQuery('.eduadmin-courselistoptions').data('fetchmonths'),
+				showvenue:			jQuery(target).data('showvenue'),
 				orderby:			jQuery('.eduadmin-courselistoptions').data('orderby'),
 				order:				jQuery('.eduadmin-courselistoptions').data('order'),
 				phrases: 	wp_edu.Phrases
@@ -137,6 +138,7 @@ edu.apiclient = {
 				template: 			jQuery(target).data('template'),
 				numberofevents:		jQuery(target).data('numberofevents'),
 				fetchmonths:		jQuery(target).data('fetchmonths'),
+				showvenue:			jQuery(target).data('showvenue'),
 				orderby:			jQuery(target).data('orderby'),
 				order:				jQuery(target).data('order'),
 				phrases: wp_edu.Phrases
@@ -166,6 +168,7 @@ edu.apiclient = {
 				eid: 			jQuery(target).data('eid'),
 				numberofevents:		jQuery(target).data('numberofevents'),
 				fetchmonths:	jQuery(target).data('fetchmonths'),
+				showvenue:			jQuery(target).data('showvenue'),
 				phrases: 		wp_edu.Phrases
 			},
 			success: function(d) {