+++ content/template/detailTemplate/template_A.php
@@ -265,6 +265,9 @@ else
 			}
 		} ?>
 	</div>
+<?php
+$showEventVenue = get_option('eduadmin-showEventVenueName', false);
+?>
 	<div class="event-table eventDays" data-eduwidget="eventlist" data-objectid="<?php echo esc_attr($selectedCourse->ObjectID); ?>"
 		data-spotsleft="<?php echo get_option('eduadmin-spotsLeft', 'exactNumbers'); ?>"
 		data-spotsettings="<?php echo get_option('eduadmin-spotsSettings', "1-5\n5-10\n10+"); ?>"
@@ -272,7 +275,9 @@ else
 		data-showmore="0"
 		data-groupbycity="<?php echo $groupByCity; ?>"
 		data-fetchmonths="<?php echo $fetchMonths; ?>"
-		<?php echo (isset($_REQUEST['eid']) ? ' data-event="' . $_REQUEST['eid'] . '"' : ''); ?>>
+		<?php echo (isset($_REQUEST['eid']) ? ' data-event="' . $_REQUEST['eid'] . '"' : ''); ?>
+		data-showvenue="<?php echo @esc_attr($showEventVenue); ?>"
+	>
 	<?php
 	$i = 0;
 	if(!empty($prices))
@@ -282,7 +287,10 @@ else
 			if($groupByCity && $lastCity != $ev->City)
 			{
 				$i = 0;
-				echo '<div class="eventSeparator">' . $ev->City . '</div>';
+				echo '<div class="eventSeparator">';
+				echo $ev->City;
+
+				echo '</div>';
 			}
 			if(isset($_REQUEST['eid']))
 			{
@@ -301,6 +309,8 @@ else
 				<div class="eventCity">
 					<?php
 					echo $ev->City;
+					if($showEventVenue && !empty($ev->AddressName))
+						echo ", " . $ev->AddressName;
 					?>
 				</div>
 				<?php } ?>