+++ includes/_options.php
@@ -72,6 +72,7 @@ function eduadmin_settings_init()
 	register_setting('eduadmin-list', 'eduadmin-showEventPrice');
 	register_setting('eduadmin-list', 'eduadmin-showCourseDays');
 	register_setting('eduadmin-list', 'eduadmin-showCourseTimes');
+	register_setting('eduadmin-list', 'eduadmin-showEventVenueName');
 
 
 	/* Global settings */