+++ includes/bookingSettings.php
@@ -94,11 +94,11 @@ else
 			<?php
 			$singlePersonBooking = get_option('eduadmin-singlePersonBooking', false);
 			?>
-			<label>
+			<!--<label>
 				<input type="checkbox" name="eduadmin-singlePersonBooking"<?php echo ($singlePersonBooking === "true" ? " checked=\"checked\"" : ""); ?> value="true" />
-				<?php _e("Participant is also customer and contact", "eduadmin"); ?>
+				<?php _e("Participant is also customer and contact (Only allow a single participant)", "eduadmin"); ?>
 			</label>
-			<br />
+			<br />-->
 			<h4><?php _e("Field order", "eduadmin"); ?></h4>
 			<?php
 			$fieldOrder = get_option('eduadmin-fieldOrder', 'contact_customer');